requirements.txt
yfinance
pandas
ta
python-telegram-bot==20.8

